refactor: merge services and serice modules
