--- client/src/modules/services/service/service.constants.js
@@ -1,10 +1,0 @@
-export default {
-  REQUEST_SERVICE : 'REQUEST_SERVICE',
-  RECEIVE_SERVICE : 'RECEIVE_SERVICE',
-  REQUEST_SAVE_SERVICE : 'REQUEST_SAVE_SERVICE',
-  SERVICE_SAVED : 'SERVICE_SAVED',
-  REQUEST_DELETE_SERVICE: 'REQUEST_DELETE_SERVICE',
-  SERVICE_DELETED: 'SERVICE_DELETED',
-  INVALID_REQUEST_SERVICE : 'INVALID_REQUEST_SERVICE',
-  NEW_SERVICE : 'NEW_SERVICE'
-};
