+++ client/src/modules/services/services.constants.js
@@ -2,5 +2,12 @@ import { generateEntitiesConstants } from '../utils/entities.js';
 
 export default {
   ...generateEntitiesConstants('services'),
-  CHANGE_FILTER: 'CHANGE_FILTER_SERVICES'
+  CHANGE_FILTER: 'CHANGE_FILTER_SERVICES',
+  REQUEST_SERVICE : 'REQUEST_SERVICE',
+  RECEIVE_SERVICE : 'RECEIVE_SERVICE',
+  REQUEST_SAVE_SERVICE : 'REQUEST_SAVE_SERVICE',
+  SERVICE_SAVED : 'SERVICE_SAVED',
+  REQUEST_DELETE_SERVICE: 'REQUEST_DELETE_SERVICE',
+  SERVICE_DELETED: 'SERVICE_DELETED',
+  INVALID_REQUEST_SERVICE : 'INVALID_REQUEST_SERVICE'
 };
